Add files via upload
This script is contains a new tool for arcGIS pro that I was developing. The script allows users to convert excel data from simple csv downloads provided by GBIF to conduct proximity statistics, and analysis of incidents within areas of interest. The script also allows someone to perform proximity analysis for points using the area_Interest parameter. You can also use this script to compare and visualize the number of points within an area to its surrounding areas. This tool would be useful for showing the distance of a species from an area of interest, such as a neighborhood or park, and compare the distance that the species is from your area of interest. I would love to receive feedback on this tool. I am currently new to using python to automate tasks, and would greatly appreciate any tips for optimizing or improving this script.