Add files via upload
This code will allow you to:
Convert simple csv files from GBIF to a geodatabase.
Project and clip the resulting feature classes.
If an area of interest and overlap area are selected and the overlap area encloses the area of interest, the proximity of points of the GBIF incidents to the area of interest will be calculated and displayed in the final product.
This script removes places where the area of interest overlaps with the overlap area layer to allow visualize of points within interest area compared to the total area given. 
If optional parameters are given, the final output layer will contain the names of these parameters.

The map attached shows how the data provided by this tool can be visualized.