Reuploaded jupyter notebook
This jupyter notebook contains the code, tables, and maps that I made using pandas and geopandas. 

This data will be part of a bigger project that I'm working on. Ultimately, the data will be used to aid in a StoryMap on ways to improve the Martin Luther King Riverside Park. As such, the data has been manipulated with this focus in mind.

The Memphis Police Department: Public Safety Incidents dataset can be obtained here: https://memphisinternal.data.socrata.com/Public-Safety/Memphis-Police-Department-Public-Safety-Incidents/ybsi-jur4

I am still very new to using pandas and geopandas for data manipulation, so if you have any feedback on ways that I could improve this script, I would appreciate it.