Add files via upload
These files are for a module and script that I made. The goal is to take user input to find locations that the user would be interested in, and then provide the option for the user to view the linear distance of the matched locations and a map of their position in reference to the locations that they're interested in.